+++ CHANGELOG.md
@@ -116,7 +116,7 @@
 - Added `craft\elements\Address`.
 - Added `craft\elements\Asset::$alt`.
 - Added `craft\elements\Asset::EVENT_AFTER_GENERATE_TRANSFORM`.
-- Added `craft\elements\Asset::EVENT_GENERATE_TRANSFORM`.
+- Added `craft\elements\Asset::EVENT_BEFORE_GENERATE_TRANSFORM`.
 - Added `craft\elements\Asset::getFs()`.
 - Added `craft\elements\Asset::setFilename()`.
 - Added `craft\elements\conditions\addresses\AddressCondition`.